104-fibonacci.c

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -12,13 +12,13 @@
 
 int main(void)
 {
-int i, j, k, j1, j2, k1, k2;
+unsigned long int i, j, k, j1, j2, k1, k2;
 j = 1;
 k = 2;
-printf("%d", j);
+printf("%lu", j);
 for (i = 1; i < 91; i++)
 {
-printf(", %d", k);
+printf(", %lu", k);
 k = k + j;
 j = k - j;
 }
@@ -28,8 +28,8 @@ k1 = k / 1000000000;
 k2 = k % 1000000000;
 for (i = 92; i < 99; ++i)
 {
-printf(", %d", k1 + (k2 / 1000000000));
-printf("%d", k2 % 1000000000);
+printf(", %lu", k1 + (k2 / 1000000000));
+printf("%lu", k2 % 1000000000);
 k1 = k1 + j1;
 j1 = k1 - j1;
 k2 = k2 + j2;
